+++ foo_ui_columns/filter_inline_edit.cpp
@@ -4,12 +4,12 @@
 namespace cui::panels::filter {
 
 bool FilterPanel::notify_before_create_inline_edit(
-    const pfc::list_base_const_t<t_size>& indices, unsigned column, bool b_source_mouse)
+    const pfc::list_base_const_t<t_size>& indices, size_t column, bool b_source_mouse)
 {
     return !m_field_data.m_use_script && !m_field_data.m_fields.empty() && column == 0 && indices.get_count() == 1
         && indices[0] != 0;
 }
-bool FilterPanel::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, unsigned column,
+bool FilterPanel::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, size_t column,
     pfc::string_base& p_text, t_size& p_flags, mmh::ComPtr<IUnknown>& pAutocompleteEntries)
 {
     t_size indices_count = indices.get_count();
